Automatically recompile outdated MIME
