--- config/config.exs
@@ -1,3 +1,0 @@
-use Mix.Config
-
-config :mime, :types, %{"video/mp2t" => "ts"}
